buildtool stuff

+++ tools/build/build.js
@@ -288,21 +288,32 @@ export const BuildDevTarget = new Juke.Target({
   ]),
   executes: async () => {
     Juke.rm('dist/.devtmp', { recursive: true })
+
+    if (fs.existsSync("dist/dev")) {
+      // Do something
+      Juke.logger.info('Only updating mods as dist/dev exists');
+
+      fs.mkdirSync("dist/dev", { recursive: true });
+      fs.cpSync('dist/modcache', 'dist/.devtmp', { recursive: true });
+      fs.cpSync('mods', 'dist/.devtmp', { recursive: true });
+      return;
+    }
+
     fs.mkdirSync("dist/dev", { recursive: true });
     fs.mkdirSync("dist/.devtmp", { recursive: true });
-    for (const folders of includeList.filter(v => !(v === "mods" || v === "config"))) {
+    for (const folders of includeList.filter(v => v !== "mods")) {
       symlinkSync(resolve(folders), resolve(`dist/dev/${folders}`));
     }
 
     // "merge" both mod folders
     fs.cpSync('dist/modcache', 'dist/.devtmp', { recursive: true });
-    fs.cpSync('mods', 'dist/.devtmp', { recursive: true, force: true });
-    fs.cpSync('dist/.devtmp', 'dist/dev/mods', { recursive: true });
-    fs.cpSync('config', 'dist/dev/config', { recursive: true });
+    fs.cpSync('mods', 'dist/.devtmp', { recursive: true });
+    symlinkSync(resolve('dist/.devtmp'), resolve('dist/dev/mods'));
+    // symlinkSync('config', 'dist/dev/config');
 
     // todo find the mod to blame, or just remove this and the filters up there if this ever gets fixed
-    Juke.logger.warn('Due to a bug with a mod, symlinking the mod folder or config causes errors which breaks game startup.')
-    Juke.logger.warn('When updating, the config and mod requires manual copy. Dev mods are packed in "dist/.devtmp"')
+    // Juke.logger.warn('Due to a bug with a mod, symlinking the mod folder or config causes errors which breaks game startup.')
+    // Juke.logger.warn('When updating, the config and mod requires manual copy. Dev mods are packed in "dist/.devtmp"')
     await packMod("dev");
   }
 })
